@@ -3,9 +3,9 @@ members = ["crates/cli", "crates/lib", "crates/macro", "e2e-tests"]
 resolver = "2"
 
 [workspace.package]
-version = "0.2.0"
+version = "0.3.0-beta.1"
 edition = "2021"
-rust-version = "1.66"
+rust-version = "1.76"
 authors = ["Alex Ostrovski <ostrovski.alex@gmail.com>"]
 license = "MIT OR Apache-2.0"
 repository = "https://github.com/slowli/externref"
@@ -22,22 +22,22 @@ quote = "1.0"
 syn = "2.0"
 tracing = "0.1.40"
 tracing-subscriber = "0.3.18"
-walrus = "0.20.3"
+walrus = "0.22.0"
 
 # Test dependencies
 assert_matches = "1.5.0"
 doc-comment = "0.3.3"
-term-transcript = { version = "0.3.0", features = ["portable-pty"] }
+term-transcript = { version = "=0.4.0-beta.1", features = ["portable-pty"] }
 test-casing = "0.1.3"
 tracing-capture = "0.1.0"
 trybuild = "1.0.99"
 version-sync = "0.9.4"
-wasmtime = "14.0.4"
+wasmtime = "25.0.1"
 wat = "1.217.0"
 
 # Internal dependencies
-externref-macro = { version = "=0.2.0", path = "crates/macro" }
-externref = { version = "=0.2.0", path = "crates/lib", default-features = false }
+externref-macro = { version = "=0.3.0-beta.1", path = "crates/macro" }
+externref = { version = "=0.3.0-beta.1", path = "crates/lib", default-features = false }
 # ^ We require an exact version in order to simplify crate evolution (e.g., to not worry
 # that future internal changes in macro implementations will break previous versions
 # of the `externref` crate).
