@@ -11,7 +11,7 @@
 
 #![recursion_limit = "128"]
 // Documentation settings.
-#![doc(html_root_url = "https://docs.rs/externref-macro/0.2.0")]
+#![doc(html_root_url = "https://docs.rs/externref-macro/0.3.0-beta.1")]
 // Linter settings.
 #![warn(missing_debug_implementations, missing_docs, bare_trait_objects)]
 #![warn(clippy::all, clippy::pedantic)]
